operator: Get ssh working on the operator side
There were a bunch of papercut bugs wth the original implementation so this
is more of a cleanup.

Also removed a leftover init container that is unnecessary.

Might also be good for us to think a bit more about how we can make these
types of scripts testable.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 39015d66e0b25d2f009c1cf1f82988505850303b
ghstack-comment-id: 3367655987
Pull-Request: https://github.com/seemethere/devserver/pull/11