operator: Split out startup script and add tests
The startup script was getting unweidly long and complicated. This split it out into a separate script and added tests to ensure it works.

This might not be the most portable for certain images but should support
most of our use cases if we need ssh.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 2111e82daad71cbf90da971236dc519001018376
ghstack-comment-id: 3367728736
Pull-Request: https://github.com/seemethere/devserver/pull/13
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>