operator: Don't mount /etc/ssh/
Mounting /etc/ssh has the weird side effect of being unable to install
the openssh package using `sudo dnf install -y openssh` due to size
limitations of our persistent volume that we mount for this.

I think this is not reasonable so instead configuring this so that we
mount at /opt/ssh and we copy things over.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 177b7ff1adb22e891a14fa67a30cc4c4f05f73e5
ghstack-comment-id: 3386285633
Pull-Request: https://github.com/seemethere/devserver/pull/44