refactor(operator): Re-organize for easier extension
I wanted to make it easier to extend to more resource types if / when we
we need to. This should re-organize the code to make it easier to extend, so
when ever we add resources we can just add a new directory with all of the things
related to that resource.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 326dc54afa74f789b13baa384a185b1d5105bd78
ghstack-comment-id: 3395387247
Pull-Request: https://github.com/seemethere/devserver/pull/57
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>