refactor(operator): Make most functions async
This is a big refactor to make the operator fully async.

The main change is to make all functions async and to use asyncio.to_thread
to run blocking operations in a separate thread.
This is a big change, but it is necessary to make the operator faster.

Test suite seems to run quite a bit faster with this change as well.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: c76429d2e4b5e615b7462a4b3669a92a592c4ec7
ghstack-comment-id: 3395419694
Pull-Request: https://github.com/seemethere/devserver/pull/58