docs: Update for factuality
I noticed some docs were not factual when I was reading them.

Went through with Gemini and cleaned this up.

Will be adding an AGENTS.md to this in my next PR to ensure
we have consistent prompts for updating docs and that they run
both a changes and factual checks.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 9e54ee756a2ed3e024c994d0f24bddea0ff65b07
ghstack-comment-id: 3395441215
Pull-Request: https://github.com/seemethere/devserver/pull/62