feat(docker): Add base image
This adds a base image for the DevServer operator.
It is a CPU-based development environment that includes `build-essential`, `curl`, and `uv` on top of a base Ubuntu LTS image.
This image is intended to be a general-purpose starting point for most development work.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: b71b223742b0b91c8b06ec7e006860b889bf9543
ghstack-comment-id: 3395472205
Pull-Request: https://github.com/seemethere/devserver/pull/64