feat(cli): Move 'ssh --proxy' to ssh-proxy
Moves the ssh proxy stuff to its own command to separate it out from our
current ssh command. This proxy seems to also not work with doing agent
forwarding so we might need to debug why this is happening.

Also added a --direct flag to the ssh command in order to allow users to
bypass the proxy. Hopefully this will be temporary and everything should
just use the proxy.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: fed50d4f84bd3f78cb1492aa5cbdd1f782cf254b
ghstack-comment-id: 3404329004
Pull-Request: https://github.com/seemethere/devserver/pull/72