chore: Clean up make targets to use uv
This was something I've been wanting to do which is just utilize uv
fully. This basically makes it so that you can run all of the things we
need without having to create a virtualenv directly and just letting uv
handle it for us.

I might add a hatch for python version at some point but for now I'm
happy to just use whatever the default is.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: f53ee6b8ed3082cb64a1de34a7336a614269f071
ghstack-comment-id: 3404512314
Pull-Request: https://github.com/seemethere/devserver/pull/73
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>