feat(operator): Add dynamic motd as part of operator
This was originally part of the dockerfile but I learned that you can
override this behavior using ForceCommand in sshd.

This allows us to have the ultimate amount of flexibility in terms of
having a login script we can depend on / modify in any container that
the user might decide to spin up.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>
ghstack-source-id: 832fdc1404f9d541ab3912ef111f8c0dbd45188d
ghstack-comment-id: 3423845804
Pull-Request: https://github.com/seemethere/devserver/pull/89
Signed-off-by: Eli Uriegas <eliuriegas@meta.com>