feat(operator): Add ability to set default flavors (#90)
This adds the ability to set default flavors so that users can run
`devctl create` and it should automatically pick up the default flavor
that your cluster admin has set up.

This will also error out if a cluster admin attempts to specify a new
flavor
as a default flavor if a default flavor is already specified.

Signed-off-by: Eli Uriegas <eliuriegas@meta.com>