feat(cli): Automatically set user's private / public keys on config creation (#110)
Now instead of hardcoding rsa for the encryption key, we now pick it up
on default config creation.

The encryption methods we look for are inspired by
[openssh](https://github.com/openssh/openssh-portable/blob/0ffb76c6590800958777cd0f7b1aaae19c74fa3f/pathnames.h#L88)

For the reviewers, please take a careful look at the changes in
`src/devservers/cli/handlers/create.py`. These were made to appease the
type checker. It made no difference in testing, however, I'm not
convinced it's correct.

---------

Co-authored-by: Sahan Paliskara <sahanpaliskara@Sahans-MacBook-Pro.local>
Co-authored-by: Eli Uriegas <1700823+seemethere@users.noreply.github.com>